:hammer: Move MIME checks into their own file
